backend/lib/manga_ocr/requirements.txt
onnxruntime>=1.16.0
pillow>=9.0.0
numpy>=1.21.0
opencv-python-headless>=4.5.0
pyclipper>=1.3.0